Refactor imports and increase error display duration
