+++ hdrop-web-next/src/api/index.ts
@@ -1,6 +1,16 @@
 export { default as ApiClient } from './ApiClient'
-export { default as Downloader } from './Downloader'
-export { default as Uploader } from './Uploader'
+
+export {
+    default as Downloader,
+    type DownloadPhase,
+    type DownloadResult
+} from './Downloader'
+
+export {
+    default as Uploader,
+    type UploadPhase,
+    type UploadResult
+} from './Uploader'
 
 export * from './request'
 export * from './response'
